upgrade to java 17 + packages

+++ src/main/java/com/zyris/calorisecalculator/dao/ProfilesRepository.java
@@ -4,5 +4,5 @@ import com.zyris.calorisecalculator.persistance.entity.DictionaryPostgreEntity;
 import com.zyris.calorisecalculator.persistance.entity.ProfilesPostgreEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProfilesRepository extends JpaRepository<ProfilesPostgreEntity, Integer> {
+public interface ProfilesRepository extends JpaRepository<ProfilesPostgreEntity, Long> {
 }
